feat!: Rename require-basline -> use-baseline
fixes #96
